refactor: separate the params to a file for write jsdoc comments
