+++ test/e2e.sh
@@ -80,46 +80,46 @@ createTmpFixture rocket-punch/basic;
 npm install;
 npm install rocket-punch@e2e --save-dev --registry "$LOCAL_REGISTRY_URL";
 npm run build;
-fileExists "$TEMP/dist/a/index.js";
-fileExists "$TEMP/dist/a/index.d.ts";
-fileExists "$TEMP/dist/b/index.js";
-fileExists "$TEMP/dist/b/index.d.ts";
-fileExists "$TEMP/dist/c/index.js";
-fileExists "$TEMP/dist/c/index.d.ts";
+fileExists "$TEMP/out/packages/a/index.js";
+fileExists "$TEMP/out/packages/a/index.d.ts";
+fileExists "$TEMP/out/packages/b/index.js";
+fileExists "$TEMP/out/packages/b/index.d.ts";
+fileExists "$TEMP/out/packages/c/index.js";
+fileExists "$TEMP/out/packages/c/index.d.ts";
 
 createTmpFixture rocket-punch/js;
 npm install;
 npm install rocket-punch@e2e --save-dev --registry "$LOCAL_REGISTRY_URL";
 npm run build;
-fileExists "$TEMP/dist/a/index.js";
-fileExists "$TEMP/dist/a/index.d.ts";
-fileExists "$TEMP/dist/b/index.js";
-fileExists "$TEMP/dist/b/index.d.ts";
-fileExists "$TEMP/dist/c/index.js";
-fileExists "$TEMP/dist/c/index.d.ts";
+fileExists "$TEMP/out/packages/a/index.js";
+fileExists "$TEMP/out/packages/a/index.d.ts";
+fileExists "$TEMP/out/packages/b/index.js";
+fileExists "$TEMP/out/packages/b/index.d.ts";
+fileExists "$TEMP/out/packages/c/index.js";
+fileExists "$TEMP/out/packages/c/index.d.ts";
 
 createTmpFixture rocket-punch/bundle;
 npm install;
 npm install rocket-punch@e2e --save-dev --registry "$LOCAL_REGISTRY_URL";
 npm run build;
-fileExists "$TEMP/dist/a/index.js";
-fileExists "$TEMP/dist/a/index.d.ts";
-fileExists "$TEMP/dist/a/icon.svg.js";
-fileExists "$TEMP/dist/a/icon.svg.d.ts";
-fileExists "$TEMP/dist/a/icon-not-bundle.svg";
-fileExists "$TEMP/dist/b/index.js";
-fileExists "$TEMP/dist/b/index.d.ts";
-fileExists "$TEMP/dist/b/test.txt.js";
-fileExists "$TEMP/dist/b/test.txt.d.ts";
-fileExists "$TEMP/dist/b/test-not-bundle.txt";
-fileExists "$TEMP/dist/c/index.js";
-fileExists "$TEMP/dist/c/index.d.ts";
-fileExists "$TEMP/dist/c/image.jpg.js";
-fileExists "$TEMP/dist/c/image.jpg.d.ts";
-fileExists "$TEMP/dist/c/image-not-bundle.jpg";
-fileExists "$TEMP/dist/c/data.yaml.js";
-fileExists "$TEMP/dist/c/data.yaml.d.ts";
-fileExists "$TEMP/dist/c/data-not-bundle.yaml";
+fileExists "$TEMP/out/packages/a/index.js";
+fileExists "$TEMP/out/packages/a/index.d.ts";
+fileExists "$TEMP/out/packages/a/icon.svg.js";
+fileExists "$TEMP/out/packages/a/icon.svg.d.ts";
+fileExists "$TEMP/out/packages/a/icon-not-bundle.svg";
+fileExists "$TEMP/out/packages/b/index.js";
+fileExists "$TEMP/out/packages/b/index.d.ts";
+fileExists "$TEMP/out/packages/b/test.txt.js";
+fileExists "$TEMP/out/packages/b/test.txt.d.ts";
+fileExists "$TEMP/out/packages/b/test-not-bundle.txt";
+fileExists "$TEMP/out/packages/c/index.js";
+fileExists "$TEMP/out/packages/c/index.d.ts";
+fileExists "$TEMP/out/packages/c/image.jpg.js";
+fileExists "$TEMP/out/packages/c/image.jpg.d.ts";
+fileExists "$TEMP/out/packages/c/image-not-bundle.jpg";
+fileExists "$TEMP/out/packages/c/data.yaml.js";
+fileExists "$TEMP/out/packages/c/data.yaml.d.ts";
+fileExists "$TEMP/out/packages/c/data-not-bundle.yaml";
 
 # EXIT
 # ==================================================----------------------------------
